requirements.txt
amqp==5.1.1
APScheduler==3.10.1
asgiref==3.7.2
beautifulsoup4==4.12.2
billiard==4.1.0
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.3
djangorestframework==3.14.0
idna==3.4
kombu==5.3.1
lxml==4.9.2
mypy-extensions==1.0.0
numpy==1.25.0
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
platformdirs==3.8.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycodestyle==2.10.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
